+++ stroom-dashboard-server/src/main/java/stroom/visualisation/VisualisationModule.java
@@ -22,23 +22,23 @@ import com.google.inject.multibindings.Multibinder;
 import stroom.entity.FindService;
 import stroom.explorer.ExplorerActionHandler;
 import stroom.importexport.ImportExportActionHandler;
-import stroom.visualisation.shared.Visualisation;
+import stroom.visualisation.shared.VisualisationDoc;
 
 public class VisualisationModule extends AbstractModule {
     @Override
     protected void configure() {
-        bind(VisualisationService.class).to(VisualisationServiceImpl.class);
+        bind(VisualisationStore.class).to(VisualisationStoreImpl.class);
 
         final Multibinder<ExplorerActionHandler> explorerActionHandlerBinder = Multibinder.newSetBinder(binder(), ExplorerActionHandler.class);
-        explorerActionHandlerBinder.addBinding().to(stroom.visualisation.VisualisationServiceImpl.class);
+        explorerActionHandlerBinder.addBinding().to(stroom.visualisation.VisualisationStoreImpl.class);
 
         final Multibinder<ImportExportActionHandler> importExportActionHandlerBinder = Multibinder.newSetBinder(binder(), ImportExportActionHandler.class);
-        importExportActionHandlerBinder.addBinding().to(stroom.visualisation.VisualisationServiceImpl.class);
+        importExportActionHandlerBinder.addBinding().to(stroom.visualisation.VisualisationStoreImpl.class);
 
         final MapBinder<String, Object> entityServiceByTypeBinder = MapBinder.newMapBinder(binder(), String.class, Object.class);
-        entityServiceByTypeBinder.addBinding(Visualisation.ENTITY_TYPE).to(stroom.visualisation.VisualisationServiceImpl.class);
+        entityServiceByTypeBinder.addBinding(VisualisationDoc.DOCUMENT_TYPE).to(stroom.visualisation.VisualisationStoreImpl.class);
 
-        final Multibinder<FindService> findServiceBinder = Multibinder.newSetBinder(binder(), FindService.class);
-        findServiceBinder.addBinding().to(stroom.visualisation.VisualisationServiceImpl.class);
+//        final Multibinder<FindService> findServiceBinder = Multibinder.newSetBinder(binder(), FindService.class);
+//        findServiceBinder.addBinding().to(stroom.visualisation.VisualisationStoreImpl.class);
     }
 }